+++ openstates/importers/organizations.py
@@ -1,4 +1,5 @@
 from django.db.models import Q
+from ._types import _JsonDict
 from .base import BaseImporter
 from ..data.models import Organization
 
@@ -7,7 +8,7 @@ class OrganizationImporter(BaseImporter):
     _type = "organization"
     model_class = Organization
 
-    def limit_spec(self, spec):
+    def limit_spec(self, spec: _JsonDict) -> _JsonDict:
         if spec.get("classification") != "party":
             spec["jurisdiction_id"] = self.jurisdiction_id